Fix building sf4e against newer versions of spdlog+fmt
`fmt` no longer automatically converts enums to ints. This isn't ideal-
I'd rather use `format_as` or something- but it's fine short-term.

+++ src/overlay/overlay.cxx
@@ -706,7 +706,7 @@ void DrawGGPOStatsOverlay(GGPOSession* ggpo, fSystem::PlayerConnectionInfo* play
 
 	GGPOErrorCode err = ggpo_get_network_stats(ggpo, players[i].handle, &stats);
 	if (!GGPO_SUCCEEDED(err)) {
-		spdlog::warn("Couldn't get GGPO stats for overlay: {}", err);
+		spdlog::warn("Couldn't get GGPO stats for overlay: {}", (int)err);
 		return;
 	}
